Work in progress, implementing various graph functions
Two new functions added to the dependency graph:

- `find_depending` which finds nodes in the graph which depend on given node
- `find_optimal` which uses this information to find the node with the `max()`
  number of inside-graph dependents (packages depending on it)

This code is far from done but adds a bunch of functionality.
All that's left to do is Build a publish strategy where the nodes with
the largest number of depending packages get published first.

At this time it's not clear how to handle circular dependencies in this graph.
If a circular dependency is detected, the program will currently run forever.

There's a few different ways of going about it:

1. Detect the circle and cancel to let the user deal with it manually.
   Later on, maybe the user could set an "ignore" file for those packages.
2. Set a range dependency which allows for fuzzy circles to be published
   (See `diesel_cli`, `diesel_migration` and `barrel` circle)

Are there other strategies not yet considered?